--- src/renderer/types.ts
@@ -1,5 +1,0 @@
-import { Item } from "../types";
-
-export type ItemWithExistance = Item & {
-    exist: boolean
-};
